Finally hopefully fix sibling search for where to create elements
The wrong subset of the tree was being searched for where to create elements. Previously only the previous siblings of the native parent would be searched. This was wrong because it skipped searching any intermediate ancestor's previous siblings, which could have also had native controls that were the sibling of interest.