Fix iOS default template experience (#76)
Xamarin.Forms on iOS seems to not like having an async MainPage load. So the default template now has an initial empty MainPage, which is a ContentPage, and the main app page in Blazor is now a ContentView (instead of a ContentPage) that is set as the content of that MainPage.