Mega refactor of special-case logic for parent/child tracking of components
Moved almost all logic of special-cased controls into the component handlers so each handler is more responsible for itself. This is done in large part by having the handlers be responsible for adding children and setting parents, thus allowing them to override methods to customize behavior for special cases such as ModalContainer, GridCell, and others.

Fixes #16